chore: move simulated browser tests to dedicated ones
