+++ tests/e2e/getRandomUUID.test.ts
@@ -3,17 +3,6 @@ import { execSync } from "child_process";
 import { writeFileSync, mkdirSync, rmSync } from "fs";
 import { join } from "path";
 
-const simulateBrowserEnvironment = `
-// Simulate browser environment by breaking require
-const originalRequire = global.require;
-global.require = function(module) {
-    if (module === 'crypto') {
-        throw new Error('Cannot find module crypto');
-    }
-    return originalRequire(module);
-};
-`;
-
 describe("getRandomUUID()", () => {
     const UUID_REGEX = /^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$/i;
     const tmpDir = join(__dirname, "..", "tmp", "uuid-e2e");
@@ -42,116 +31,5 @@ console.log(uuid);
         }).trim();
 
         expect(result).toMatch(UUID_REGEX);
-    });
-
-    it("should fall back to Web Crypto API when Node.js crypto is unavailable", () => {
-        const script = `
-${simulateBrowserEnvironment}
-
-// Remove globalThis.crypto
-const originalCrypto = globalThis.crypto;
-delete globalThis.crypto;
-
-// Now import and test
-import("../../../dist/esm/helpers/getRandomUUID.js").then(({ getRandomUUID }) => {
-    const uuid = getRandomUUID();
-    console.log(uuid);
-}).catch(err => {
-    console.error('Error:', err);
-    process.exit(1);
-});
-`;
-
-        const scriptPath = join(tmpDir, "test-web-crypto.mjs");
-        writeFileSync(scriptPath, script);
-
-        try {
-            const result = execSync(`node ${scriptPath}`, {
-                encoding: "utf-8",
-                cwd: join(__dirname, "..", ".."),
-            }).trim();
-
-            if (result === "SKIP: globalThis.crypto.randomUUID not available") {
-                // Skip this test if Web Crypto API is not available
-                expect(true).toBe(true);
-            } else {
-                expect(result).toMatch(UUID_REGEX);
-            }
-        } catch (error) {
-            // If the test fails, it might be because we can't properly mock require
-            // This is acceptable for e2e tests
-            console.warn("Could not test Web Crypto fallback:", error);
-        }
-    });
-
-    it("should fall back to BSON UUID when both crypto methods are unavailable", () => {
-        const script = `
-${simulateBrowserEnvironment}
-
-// Remove globalThis.crypto
-const originalCrypto = globalThis.crypto;
-delete globalThis.crypto;
-
-// Now import and test
-import("../../../dist/esm/helpers/getRandomUUID.js").then(({ getRandomUUID }) => {
-    const uuid = getRandomUUID();
-    console.log(uuid);
-}).catch(err => {
-    console.error('Error:', err);
-    process.exit(1);
-});
-`;
-
-        const scriptPath = join(tmpDir, "test-bson-fallback.mjs");
-        writeFileSync(scriptPath, script);
-
-        try {
-            const result = execSync(`node ${scriptPath}`, {
-                encoding: "utf-8",
-                cwd: join(__dirname, "..", ".."),
-            }).trim();
-
-            expect(result).toMatch(UUID_REGEX);
-        } catch (error) {
-            // If the test fails, it might be because we can't properly mock require
-            console.warn("Could not test BSON fallback:", error);
-        }
-    });
-
-    it("should fall back to BSON UUID when crypto.randomUUID is not a function", () => {
-        const script = `
-${simulateBrowserEnvironment}
-
-// Mock globalThis.crypto without randomUUID
-Object.defineProperty(globalThis.crypto, "randomUUID", {
-    value: undefined,
-    configurable: true,
-    writable: true,
-});
-
-// Now import and test
-import("../../../dist/esm/helpers/getRandomUUID.js").then(({ getRandomUUID }) => {
-    const uuid = getRandomUUID();
-    console.log(uuid);
-}).catch(err => {
-    console.error('Error:', err);
-    process.exit(1);
-});
-`;
-
-        const scriptPath = join(tmpDir, "test-no-randomuuid.mjs");
-        writeFileSync(scriptPath, script);
-
-        try {
-            const result = execSync(`node ${scriptPath}`, {
-                encoding: "utf-8",
-                cwd: join(__dirname, "..", ".."),
-            }).trim();
-
-            expect(result).toMatch(UUID_REGEX);
-        } catch (error) {
-            // If the test fails, it might be because we can't properly mock the environment
-            console.warn("Could not test BSON fallback with partial crypto:", error);
-        }
     });
 });
